movies-microservice/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
black==22.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
cryptography==38.0.4
dependency-injector==4.40.0
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djwto @ git+https://github.com/hr-curtissimo/djwto.git@a9089010b50d2e818a478bd9a678b83d8eb666fd
factory-boy==3.2.1
Faker==15.3.4
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
Markdown==3.4.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.9
whitenoise==5.3.0
wrapt==1.14.1
